backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
alembic==1.14.1

# LLM Providers
openai==1.59.3
httpx==0.28.1

# Document Processing
PyMuPDF==1.25.1
python-docx==1.1.2
tiktoken==0.8.0

# File Upload
python-multipart==0.0.18

# Utilities
python-dotenv==1.0.1
